Removed Sumariser component from Plant.
