+++ Prototypes/WheatPMFPrototype/WheatPMFtest.apsimx
@@ -1370,8 +1370,7 @@
         <FractionRemoved>0</FractionRemoved>
         <FractionToResidue>0</FractionToResidue>
       </GenericAboveGroundOrgan>
-      <Summariser>
-        <Name>Summariser</Name>
+
         <CompositeBiomass>
           <Name>AboveGround</Name>
           <Propertys>
@@ -1427,7 +1426,7 @@
             <string>[Grain].Dead</string>
           </Propertys>
         </CompositeBiomass>
-      </Summariser>
+
       <CropType>Wheat</CropType>
     </Plant>
   </Replacements>
@@ -2731,7 +2730,7 @@ namespace Models
         <Line>Solid</Line>
         <TableName>Report</TableName>
         <XFieldName>Clock.Today</XFieldName>
-        <YFieldName>WheatPMFPrototype.Summariser.AboveGround.Wt</YFieldName>
+        <YFieldName>WheatPMFPrototype.AboveGround.Wt</YFieldName>
         <ShowInLegend>false</ShowInLegend>
         <Cumulative>false</Cumulative>
         <CumulativeX>false</CumulativeX>
@@ -2749,7 +2748,7 @@ namespace Models
         <Line>Solid</Line>
         <TableName>Report</TableName>
         <XFieldName>Clock.Today</XFieldName>
-        <YFieldName>WheatPMFPrototype.Summariser.BelowGround.Wt</YFieldName>
+        <YFieldName>WheatPMFPrototype.BelowGround.Wt</YFieldName>
         <ShowInLegend>false</ShowInLegend>
         <Cumulative>false</Cumulative>
         <CumulativeX>false</CumulativeX>
@@ -2767,7 +2766,7 @@ namespace Models
         <Line>Solid</Line>
         <TableName>Report</TableName>
         <XFieldName>Clock.Today</XFieldName>
-        <YFieldName>WheatPMFPrototype.Summariser.Total.Wt</YFieldName>
+        <YFieldName>WheatPMFPrototype.Total.Wt</YFieldName>
         <ShowInLegend>false</ShowInLegend>
         <Cumulative>false</Cumulative>
         <CumulativeX>false</CumulativeX>
@@ -2950,7 +2949,7 @@ namespace Models
         <Line>Solid</Line>
         <TableName>Report</TableName>
         <XFieldName>Clock.Today</XFieldName>
-        <YFieldName>WheatPMFPrototype.Summariser.AboveGround.N</YFieldName>
+        <YFieldName>WheatPMFPrototype.AboveGround.N</YFieldName>
         <ShowInLegend>false</ShowInLegend>
         <Cumulative>false</Cumulative>
         <CumulativeX>false</CumulativeX>
@@ -2968,7 +2967,7 @@ namespace Models
         <Line>Solid</Line>
         <TableName>Report</TableName>
         <XFieldName>Clock.Today</XFieldName>
-        <YFieldName>WheatPMFPrototype.Summariser.BelowGround.N</YFieldName>
+        <YFieldName>WheatPMFPrototype.BelowGround.N</YFieldName>
         <ShowInLegend>false</ShowInLegend>
         <Cumulative>false</Cumulative>
         <CumulativeX>false</CumulativeX>
@@ -2986,7 +2985,7 @@ namespace Models
         <Line>Solid</Line>
         <TableName>Report</TableName>
         <XFieldName>Clock.Today</XFieldName>
-        <YFieldName>WheatPMFPrototype.Summariser.Total.N</YFieldName>
+        <YFieldName>WheatPMFPrototype.Total.N</YFieldName>
         <ShowInLegend>false</ShowInLegend>
         <Cumulative>false</Cumulative>
         <CumulativeX>false</CumulativeX>
@@ -3004,7 +3003,7 @@ namespace Models
         <Line>Solid</Line>
         <TableName>Report</TableName>
         <XFieldName>Clock.Today</XFieldName>
-        <YFieldName>WheatPMFPrototype.Summariser.TotalLive.N</YFieldName>
+        <YFieldName>WheatPMFPrototype.TotalLive.N</YFieldName>
         <ShowInLegend>false</ShowInLegend>
         <Cumulative>false</Cumulative>
         <CumulativeX>false</CumulativeX>
